+++ src/DocumentFormat.OpenXml/Packaging/ImagePartExtensions.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using DocumentFormat.OpenXml.Features;
 using System;
 
 namespace DocumentFormat.OpenXml.Packaging
@@ -17,20 +18,19 @@ namespace DocumentFormat.OpenXml.Packaging
         /// </summary>
         /// <param name="main">The source MainDocumentPart that is adding the ImagePart.</param>
         /// <param name="extension">The file extension of the image.</param>
-        /// <see cref="MainDocumentPart.AddImagePart(ImagePartType)" />
         public static ImagePart AddImagePart(this MainDocumentPart main, string extension)
         {
             if (main is null)
             {
                 throw new ArgumentNullException(nameof(main));
             }
-
             if (string.IsNullOrEmpty(extension))
             {
                 throw new ArgumentException($"'{nameof(extension)}' cannot be null or empty.", nameof(extension));
             }
 
-            return main.AddImagePart(ImagePartTypeInfo.GetImagePartType(extension));
+            // return main.AddImagePart(ImagePartTypeInfo.GetImagePartType(extension));
+            return (ImagePart)OpenXmlPartExtensions.InitPart(main, new ImagePart(), ImagePartTypeInfo.GetImagePartType(extension));
         }
     }
 }