+++ test/DocumentFormat.OpenXml.Tests/PptxTests01.cs
@@ -407,7 +407,7 @@ namespace DocumentFormat.OpenXml.Tests
 
                 var tnPart = doc.AddThumbnailPart(ThumbnailPartType.Jpeg);
                 doc.DeletePart(tnPart);
-                tnPart = doc.AddThumbnailPart("image/jpg");
+                tnPart = doc.AddThumbnailPart(ThumbnailPartType.Jpeg, "image/jpg");
 
                 var v = new OpenXmlValidator(FileFormatVersions.Office2013);